requirements.txt
python-dotenv
pandas
numpy 
requests
flask
pymongo
certifi
pymongo[srv]==3.12
scikit-learn
pyaml
# -e .